requirements.txt
tensorflow==2.5.0
pillow==8.2.0
streamlit==0.82.0
